requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.2.10
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==3.0
defusedxml==0.7.0rc1
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.42.0
django-crispy-forms==1.9.0
django-embed-video==1.3.3
django-environ==0.4.5
django-filter==2.3.0
django-shortcuts==1.6
django-suit==2.0a1
filelock==3.0.12
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.19
oauthlib==3.1.0
packaging==20.4
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.6
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.7
# pypiwin32==223
python3-openid==3.2.0
pytz==2019.3
# pywin32==228
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.0
txaio==20.4.1
urllib3==1.25.9
virtualenv==20.0.17
webencodings==0.5.1
# win10toast==0.9
wrapt==1.11.2
zope.interface==5.1.0
